Inventory and warehouse loading fix
Changed the way how inventory data is send to client to send in packets of 10 items to prevent the client from crashing from getting too much data at once.
This also enables to properly add the warehouse data.
Inventory size is still limited to a max of 130 on the client-side for the 1.0 client.