[Game] fix: Potential or partial fix for disconnects from raid
Changed code so Character no longer release their ObjId, but insteads links it to the Character.Id in a dictionary. That way as long as the server is running, the server will always assign the same ObjId to the same character, and this way can't break the raid/party window's selections.
Still needs more finetuning related to offline status display when moving members around.