[Game] fix: Party/Raid disband behaviour and other small fixes
Fixed when a party or raid should disband on it's own.
Added support to SCChatMessagePacket to automatically add a GM tag to names when the "gmFlag" attribute has been set on the account.
Made it so TImeManager actually closes when you close the server and doesn't leave the console hanging.
Updated some text of GM commands
Updates /testguild so it can take spaces in the guildname