[Game] add: Chat Manager WIP
Fixed a few things with guild chat.
Fixed a few things with guilds themselves.
Fixed the Patron/AH buff we hacked in to only apply to characters instead of all unit types.
Added /party /raid and /cmd chat
Changed code so Character no longer release their ObjId, but insteads links it to the Character.Id in a dictionary. That way as long as the server is running, the server will always assign the same ObjId to the same character, and this way can't break the raid/party window's selections.
Fixed Party/Raid disband behaviour and other small fixes
Fixed when a party or raid should disband on it's own.
Added support to SCChatMessagePacket to automatically add a GM tag to names when the "gmFlag" attribute has been set on the account.
Made it so TImeManager actually closes when you close the server and doesn't leave the console hanging.
Updated some text of GM commands
Updates /testguild so it can take spaces in the guildname
Code Cleanup and comments added for ChatManager