[Game] WIP: Transform
Removed WorldPosition and WorldRotation from Transform and replaced all instances to World.Position and World.Rotation.
Hopping on and off of objects (vehicles) now properly sets the parent transforms.

Transform.GetWorldPosition() is not implemented correctly yet.
Temporary changed /move ??? ToMe so it only shows the expected coordinates instead of actually teleporting. This is to test Transform.World behaviour while standing on a vehicle. The 3D Math is all wrong btw, resulting in objects having the wrong location/rotation when it has a parent Transform.
World will just return the local transform if no parent is set.