[Game] WIP: Housing - Furniture
Added proper saving and loading or furniture positions and rotations.
Furniture parenting is currently not working, so this is technically a bug.

Known issue, things like "Wooden Chest" (item 18483) cannot be recovered for some reason yet after a server restart. This is likely related to ItemIds or owner information on the doodad.