[Game] Housing permissions update
Things placed on owned land will now be protected by a fail-safe, even if it is still visually off the land with the current bug related to rotation of the land.
Changed the permissions property to take into account the always public flag of property, even if the database said it's private.