[Game] Fix: Drowning
Uncommented a SetPosition in Transform that was preventing drowning (and likely other things like zone borders) to take effect.
I'm sure there was a reason this was commented in the first place by me, but I can't remember why this was.
So uncommented for now, and let it probably break some other stuff ^_^