[Game] WIP: Ucc, SpecialEffectType, Backpacks
Removed the Id that was set on slaves when spawning, it prevented using some of the child doodads.
Doodads now properly spawn with their Ucc
Simplified SCItemTaskSuccessPacket so that you can directly call it with a task without having to create a array, and you can now also just call it with null for the forceRemove pamareter.
Added and Implemented SCItemUccDataChangedPacket
Implemented CleanupUccEffect
Added TryAddNewItem to Inventory, you can use this as a simplified Bag.AquireDefaultItem, but it takes into account of equipping backpacks directly on the character if available.
Replaced the magic numbers in GainLootPackItemEffect
Added SpecialEffectActionType field to SpecialEffectAction (makes it easier to implement missing actions)
Made GiveBmMileage actually change the BmPoints (Loyalty?)
Changed how items are consumed in Skill.ApplyEffects, they are now consumed after the skill.
Updated GM command /additem to properly equip all types on backpacks, not just normal tradepacks.
Various comments updated
Removed SlaveCall.cs and PetCall.cs in Skill/Effects/SpecialEffects as these names are not use, and are likely left over from older namings.