[Game] UCC saving into DB
Changed uccs's data field to MediumBlob so it can store the .dds file inside of it (normal blob is limited to 64kB, MediumBlob gets around this)
I left in the option to still store them in the UserData folder by enabling #define STORE_UCC_AS_FILE in UccManager.cs if you really want to. This was used during debugging.
Please note that it's not supported to change this setting on a already had UCCs stored before (it will not migrate), which I why I'm setting it as a compiler define instead of a configuration.

Added locking to _downloadQueue where it wasn't before to get rid of the warning. Also fixed the lock object naming.

General code cleanup