[Game] Configuration Manager rewrite
Removed most of the individual configuration management system that were in place in favor of the one we used for the main config.json file.
The old AppConfiguration functionality has been expanded to include the other removed ones as their own section in the config.
Fragmented configurations can be placed in the Configurations folder and will be added to the config building in alphabetical order with the main config.json going first, and adding command-line arguments at the end.

[Game] Logout Message
Implemented quit game message.
Also implemented that it takes 5 minutes to log out if you are still in combat