[Game] WIP: Reading Client World Data
Removed the need for data/worlds.json and data/*/zones.json, this data is now directly read from the clientdata instead.
Added ZoneKey to Sectors (Region class) that replaces the functionality of World.ZoneIds, this is done because it was just easier to implement that way when reading client data.
Removed the XmlWorldZoneList class as it doesn't really warrant to be there to begin with, and shorted the names of related Xml reading classes.
Updated GetZoneOriginCell to use client data instead of hard-coded