[Game] Client Integration - Heightmaps
Added a "thread-safe" version of AAPak.ExportFileAsStream(Cloned)
Added code and classes to load heightmaps from client dat. (originally by Mr.Nikes)
Added a PreferClientHeightMap setting to ClientData Configuration that can be used to try and force client data to be loaded first instead of the pre-generated heightmaps.
Both are supported, and both will try to fall-back to the other.
Added and edited some of the magic numbers in WorldManager.
Moved the loading of ZoneManager, WorldManager and Heightmaps a bit earlier in the loading chain to reduce waiting times before the data is required to spawn units.
Added a Info line to the logs in case it needs to wait.
Moved some logging from Debug to Trace