[All] Consolidated MySQL.cs into Commons
Removed the individual instances of MySQL in AAEmu.Game and AAEmu.Login and moved them to AAEmu.Commons instead.
Create a separate file for MySqlConnectionSettings so it is no longer service dependent.
Updated all namespace references as needed.
Did a small bit of code cleanup while in the modified files.