[Game] Charged Equipment
Added support to apply buffs of equipment with timed buffs (mostly quest items and cosmetics)
Do note that if you have previously timed or charged items, then these will not work as intended until re-acquired.

Items with a charge counter should work in theory, but since no such items exist in 1.2, I am unable to verify this. Expiration of such items are also not fully implemented when the last charge would be used, as we have no way to actually test this, and should be implemented on the UseItem section of the code.