[Game] Items with expiration date and limited buff effect times (#486)
* [Game] WIP: Item Expiration

Added base fields for working with item expiration functionality
Added base timer
Basic item expiration should now work, but timers are not saved to DB yet, so they reset if you restart the server.
Added DB entries for items.
Charges not implemented yet

* [Game] Charged Equipment

Added support to apply buffs of equipment with timed buffs (mostly quest items and cosmetics)
Do note that if you have previously timed or charged items, then these will not work as intended until re-acquired.

Items with a charge counter should work in theory, but since no such items exist in 1.2, I am unable to verify this. Expiration of such items are also not fully implemented when the last charge would be used, as we have no way to actually test this, and should be implemented on the UseItem section of the code.

Also move where UpdateGearBonuses() is applied in the code to the newly added OnEnterContainer and OnLeaveContainer of the EquipmentContainer class.

* CodeFactor stuff