[Game] WIP: Mail quest rewards
Changed QuestsActSupplyItem and QuestActSupplySelectiveItem so they don't directly give the items, but rather add their templates to a reward pool.
If everything goes through the component correctly, only then will it actually create the items.
If there is enough room, they will be pushed to the inventory in the usual fashion, otherwise they will all be send as mail(s).

Mail format is not final yet.