[Game] Fixed Multi-craft (#494)
* [Game] Fixed Multi-craft

Removed a invalid check of IsPack that isn't doing what it's expected to do.
Changed it so that repeat crafts start using a Scheduled Task taking into account cooldowns.
The problem was mostly trying to repeat crafts but getting cancelled because of still being in the cooldown period.

* [Game] Fixed Multi-craft - CodeFactor

For some reason I always seem to overlook at least one blank line.
(╯‵□′)╯︵┻━┻