[Game] Fix MoveType bug
When grabbing onto or jumping of things like a ladder, the client will not always send the ObjId, but send a value of 1 instead. This will cause it to try and link to whatever was the first object spawned (which is usually somewhere else anyway). This value needs to be ignored.
For safety, I also changed the minimum Id to 0x100 in the ObjectIdManager.
This fix is mostly to fix the ladder issue on boats, where it would glitch/phase if you try to grab a ladder while already on a boat.
This may also fix some other parenting issues, and it may also break some other stuff, but I couldn't find anything that wasn't broken before.