[Game] Added water editor (#499)
* [Game] Added GM commands for water 

Modified to the ScriptsCompiler to give slightly more detailed information on the more exotic errors.
Modified /fly to be able to be used as a toggle
Added **friendly** as a option to ``/setfaction``
Added JSON settings to the water body classes
Added ``/wateredit`` (alias ``/wedit``)  to allow editing of water bodies inside the world. Run without arguments to get a list of subcommands.
Added lock() when working with water bodies to try and prevent crashes when editing water
Modified WorldManager so it prefers loading the custom saved json data over the estimated client data.

* Code-Factor

* [Game] /wateredit movepoint

Changed /wateredit so that you can use movepoint (mp) without providing a index as well. When index is omitted, it will move the next point instead.