[Test] Updated Tests related to GM Command Parameters
Changed some of the test data since it can now support double type.
Changed Parameter code so that all floats/doubles are handle in a culture invariant way.
Updated IdManager to be able to Force-Re-Initialize the Instance. This is needed to fix batch test runs (in Rider) since the object was not created anew for every test.