[Login] Allow hostnames and local connections (#548)
Changed the Login server so it can also accept a hostname for the server name field in the DB. It will try to resolve the name using a DNS lookup and then just locally store the IP for use on connections.

Changed ACWorldCookiePacket so it will use the same IP as used to connect with if the client is on the same machine as the login server.
This is a debugging feature that allows you to run a local login + game server using a local client, while still allow others to connect using your external IP provided in the DB.
This requires the port binding is be set to the default asterisk (*), and assumes the game server is running on the same machine as the login server.