[Game] Fix: Trade pack Ratio and Overburdened (#707)
* [Game] Fix: Trade pack Ratio and Overburdened

Trade pack ratios were not updated as expected. It would keep lowering the % based on the total amount of packs set as the pack counter never got reset.

* [Game] Add Specialty configuration

Added code to add SpecialtyManager's settings to the Configurations Folder.
Internally changed the price rates to double instead of int to fix a bug where 1 pack / minute (tick) could never decrease the % for that pack type.