[Game] Fix: CPU Usage
Changed some things to help optimize performance.
Changed Physics Tick Per Second to 15 instead of 60
Moved some of the common packet logs to Trace instead of Debug
Changed ActiveRegionTick so it only checks characters instead of all regions since there is not region specific code in it.
Changed FindPath Info so that it only shows if it took a long time.
Added a if FileExists check to loading Simulation paths to avoid creating a unneeded exception there.