[Game] Fix quest starting items
Not sure how the original devs intended to use the Quest Starting items, but a whole bunch of them link to skill 11141 which is invalid. I added a workaround for any item with impl_id 10 (quest statrers) to use a alternative lookup method to find the real required QuestId. Code is very ugly, but does work.
Also added a HasQuestCompleted function, and added a bool result to Add quest to return if it actually added it.