[Game] Rewrite SendMessage to have arguments removed (#824)
* [Game] Fix typo in QuestState

Fixed a typo in ``QuestStartState.Start`` for ``QuestActObjSphere`` if there wasn't a related quest sphere found. The error message send to the client was formatted wrong in this case.
This prevented players from accepting the arena related quests.

* [Game] Replaced SendMessage

- Changed ``SendMessage()`` so that it only has two variants anymore, removing the option to use the parameters argument in favor of preformatted strings.
- ``SendMessage(string)`` and ``SendMessage(ChatType, string, Color?);`` are now the only options.
- Updated all instances that needed a change to the new format.
- Update Tests to reflect this as well.

* [Test] Commented out test that now fail

I honestly got no idea how to fix these with the changes to ``SendMessage()``, so for now we're commenting them out.