[Login] Added option to skip the host resolving (#966)
Added a ``SkipHostResolve`` option to ``Config.json`` to allow the login server to skip host name resolving and simply processing the raw IP used in the ``game_servers`` table.
This can be useful for making specific setups work correctly. Recommended to keep this setting ``false`` unless it is consistently resolving to the wrong IP when starting.