[Game] Workaround for ship skills targetting more than the ship
Added a check for source and siege target properties of skill to single out skill of ships that need to apply buffs to itself.
We are likely missing some other way where skill targeting is not producing the intended results. But `siege_target` seems to be the only property where this is different from other skills with radius set.