[Data] Fix Glitterstone Mine enemy npc coordinates
A follow-up of my previous fix, this one fixes the spawn coordinates of the enemy npcs in this area so that they spawn at the correct elevations.

This fix specifically fixes all instances of  sluggish seabug including the 3 at the Caenord docks for some reason, all instances of bone grub, and Grave Maw (boss grub).